Add files via upload
Explore the intricacies of Covid-19 data across the African continent through my dashboard. 

Key Features of the Dashboard: 

- Comprehensive Data Overview: Instantly grasp the gravity of the situation with total case numbers, death tolls, and mortality rates at a glance. 
- Country-Specific Insights: Delve into the impact on individual countries with a detailed bar chart depicting death counts, offering a comparative analysis across the continent. 
- Geospatial Analysis: Navigate the geographic spread and intensity of the health impact with a color-coded map, allowing for an at-a-glance understanding of regional variances. 
- Temporal Trends: Observe the progression of the event over time with a line graph that traces infection percentages, illustrating the dynamic nature of the health landscape.

This dashboard is a product of rigorous data analysis and a keen eye for design, ensuring that complex information is accessible and actionable. Created with Tableau, it reflects a blend of analytical precision and creative visualization techniques.  Immerse Yourself in Data: For an interactive experience and to explore the data more deeply, visit the Tableau Public visualization. (https://public.tableau.com/shared/PH6MZJFDG?:display_count=n&:origin=viz_share_link)  Perfect for public health professionals, researchers, and anyone with an interest in health data analytics, this dashboard is a cornerstone of my data visualization portfolio. It represents my commitment to transforming data into meaningful stories and actionable insights.